requirements.txt
requests>=2.32.0
lxml>=5.2.2
beautifulsoup4>=4.12.3
html5lib>=1.1
mechanize>=0.4.10
python-dotenv>=1.0.1
typer>=0.12.5
rich>=13.8.1
pyinstaller>=6.10.0